studai-backend/requirements.txt
fastapi
python-dotenv
uvicorn
moviepy
pillow
requests
openai
edge_tts
pillow
pydantic
pypdf
azure-storage-blob
python-multipart
PyPDF2
assemblyai
asyncio
pydub
faster-whisper
PyMuPDF
pytesseract
httpx
azure-cognitiveservices-speech